pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "python-owasp-zap-v2-4>=0.1.0",
    "reportlab>=4.4.3",
    "requests>=2.32.4",
]
